Add syscall structs and docs
