+++ src/rtc.h
@@ -8,15 +8,7 @@ See the file LICENSE for details.
 #define RTC_H
 
 #include "kerneltypes.h"
-
-struct rtc_time {
-    uint8_t second;
-    uint8_t minute;
-    uint8_t hour;
-    uint8_t day;
-    uint8_t month;
-    uint16_t year;
-};
+#include "sys_rtc_struct.h"
 
 void rtc_init();
 void rtc_read(struct rtc_time *t);
