+++ src/sys_clock_struct.h
@@ -1,0 +1,16 @@
+/*
+Copyright (C) 2016 The University of Notre Dame
+This software is distributed under the GNU General Public License.
+See the file LICENSE for details.
+*/
+
+#ifndef SYS_CLOCK_STRUCT_H
+#define SYS_CLOCK_STRUCT_H
+
+typedef struct {
+    uint32_t seconds;
+    uint32_t millis;
+} clock_t;
+
+#endif
+
