+++ src/sys_rtc_struct.h
@@ -1,0 +1,20 @@
+/*
+Copyright (C) 2016 The University of Notre Dame
+This software is distributed under the GNU General Public License.
+See the file LICENSE for details.
+*/
+
+#ifndef SYS_RTC_STRUCT_H
+#define SYS_RTC_STRUCT_H
+
+struct rtc_time {
+    uint8_t second;
+    uint8_t minute;
+    uint8_t hour;
+    uint8_t day;
+    uint8_t month;
+    uint16_t year;
+};
+
+#endif
+
