+++ src/syscall.h
@@ -12,7 +12,6 @@ See the file LICENSE for details.
 #define SYSCALL_exit     1
 #define SYSCALL_testcall 2
 #define SYSCALL_yield    3
-// TODO: change these syscall numbers
 #define SYSCALL_clock_read 4
 #define SYSCALL_sleep 5
 #define SYSCALL_rtc_read 6
